requirements.txt
streamlit
numpy
sympy
matplotlib
google-generativeai
Pillow
